Final Commit
Finalizing everything. I almost made it look like it was supposed to, but with some visual glitches on the GitHub website as well as not being able to merge when I needed to prevented some things from going to plan. This final push should have all the information on STEPS.